add : searching, pagination; remaining : sorting

+++ src/app/modules/user/list-container/list-container.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { Observable } from 'rxjs';
+import { Observable, of } from 'rxjs';
 import { User } from 'src/app/core/model/user.model';
 import { UserService } from 'src/app/shared/services/user-services/user.service';
 
@@ -10,7 +10,7 @@ import { UserService } from 'src/app/shared/services/user-services/user.service'
 })
 export class ListContainerComponent implements OnInit {
 
-  public user$ : Observable<User>;
+  public user$ : Observable<User> = of();
 
   constructor(private userService : UserService) { }
 
@@ -25,7 +25,10 @@ export class ListContainerComponent implements OnInit {
   }
 
   removeUser(event){
-    this.userService.deleteData(event);
+    if(confirm('Are you sure ?')){
+      this.userService.deleteData(event);
+    }
+    
   }
 
 }
